requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-google-genai
langchain-huggingface
langchain-community
google-generativeai
faiss-cpu
sentence-transformers
PyYAML